TIP-874: move EntityWithFamilyVariantAttributesProvider
